feat(pre-selection): SelectComplete and SelectNonExpiring
