+++ src/skfolio/pre_selection/__init__.py
@@ -1,7 +1,13 @@
-from skfolio.pre_selection._pre_selection import (
-    DropCorrelated,
-    SelectKExtremes,
-    SelectNonDominated,
-)
+from skfolio.pre_selection._drop_correlated import DropCorrelated
+from skfolio.pre_selection._select_complete import SelectComplete
+from skfolio.pre_selection._select_k_extremes import SelectKExtremes
+from skfolio.pre_selection._select_non_dominated import SelectNonDominated
+from skfolio.pre_selection._select_non_expiring import SelectNonExpiring
 
-__all__ = ["DropCorrelated", "SelectKExtremes", "SelectNonDominated"]
+__all__ = [
+    "DropCorrelated",
+    "SelectKExtremes",
+    "SelectNonDominated",
+    "SelectComplete",
+    "SelectNonExpiring",
+]
